requirements.txt
# LINE Bot 相關
line-bot-sdk==2.4.1
flask==2.2.3
python-dotenv==1.0.0

# 資料庫相關
# sqlite3>=3.35.0

# 日期時間處理
pytz==2022.7.1
python-dateutil==2.8.2
schedule==1.1.0

# 自然語言處理
requests>=2.26.0

# Web 介面
Flask-WTF>=0.15.1
Flask-Login>=0.5.0
Flask-SQLAlchemy>=2.5.1

# 數據視覺化
matplotlib>=3.4.3
pandas>=1.3.3

# 工具類
pydantic>=1.9.0
loguru>=0.5.3

# 部署相關
gunicorn==20.1.0
whitenoise>=5.3.0
python-json-logger>=2.0.2

# 監控相關
prometheus-flask-exporter>=0.18.2
sentry-sdk>=1.5.0

# 新增的依賴
sqlalchemy==2.0.7
apscheduler==3.10.1
openai==0.27.2
pinecone-client==2.2.1
tqdm==4.65.0
torch==2.0.0
transformers==4.28.1
numpy==1.24.2
scipy==1.10.1
langchain==0.0.142
langchain-openai==0.0.2.post1

# HTML轉圖像相關依賴
selenium==4.9.1
Pillow==9.5.0
webdriver-manager==3.8.6 